Refactor to feature-based architecture
Reorganize project into feature and shared modules and migrate auth/home code. Key changes:

- Introduced ARCHITECTURE.md describing the feature-based module layout and path aliases.
- Reorganized source: moved UI/shared code to src/shared/* and feature-specific code to src/features/* (auth, home). Added feature index exports (features/auth, features/home).
- Auth feature: added AuthCard, migrated FormInput, moved/renamed hooks to useLogin/useRegister, introduced types and validation schemas (auth.schema.ts), and updated tests accordingly. Hooks now use isLoading and return typed errors.
- i18n: split message files into i18n/messages/{auth,home}/{en,es}. Updated i18n/request.ts to dynamically import and merge feature message modules per locale.
- Layout & routes: added (auth) route group and moved login/register pages and auth layout. Updated app layout to use shared header and providers.
- Shared utilities & components: moved utils, icons, hooks (useScrollAnimation), UI components, and tests into shared/ and updated imports across the codebase. Minor UI tweaks (banner opacity, floating shapes, etc.).
- README updated to use repository URL and npm ci; tsconfig.json modified.

This commit is a large refactor focused on modularizing the codebase for better feature ownership and reuse; no functional API changes are intended beyond reorganizing and small UX/implementation adjustments.
