+++ i18n/messages/home/en.json
@@ -1,0 +1,55 @@
+{
+  "nav": {
+    "home": "Home",
+    "leaderboard": "Leaderboard",
+    "login": "Login",
+    "register": "Register"
+  },
+  "home": {
+    "badge": "World Cup 2026 Predictions",
+    "hero": {
+      "title": "Predict. Compete.",
+      "titleHighlight": "Win.",
+      "subtitle": "Join the ultimate World Cup prediction game. Pick match scores, compete with friends, and climb the leaderboard to become the ultimate football oracle."
+    },
+    "cta": {
+      "getStarted": "Get Started",
+      "signIn": "Sign In"
+    },
+    "howItWorks": {
+      "title": "How It Works",
+      "subtitle": "Simple, fun, and competitive. Start predicting in minutes.",
+      "steps": {
+        "pickScores": {
+          "title": "Pick Your Scores",
+          "description": "Predict the exact score of every match before the opening whistle blows."
+        },
+        "earnPoints": {
+          "title": "Earn Points",
+          "description": "Get 3 points for exact scores and 1 point for correctly predicting the winner or a draw."
+        },
+        "compete": {
+          "title": "Global Competition",
+          "description": "Test your knowledge against users worldwide and climb the global leaderboard."
+        },
+        "joinLeague": {
+          "title": "Private Leagues",
+          "description": "Create or join leagues with friends and family to compete for ultimate bragging rights."
+        },
+        "matchLock": {
+          "title": "Prediction Lock",
+          "description": "Predictions lock 5 minutes before kick-off. Make sure to save your picks in time!"
+        },
+        "masterTheBracket": {
+          "title": "Master the Bracket",
+          "description": "As the tournament progresses, fill out the knockout stages all the way to the final."
+        }
+      }
+    },
+    "finalCta": {
+      "title": "Ready to Start Predicting?",
+      "subtitle": "Join thousands of football fans competing to be the ultimate World Cup oracle.",
+      "button": "Create Your Account"
+    }
+  }
+}
