--- src/components/home/index.ts
@@ -1,3 +1,0 @@
-export { BannerSection } from "./banner-section";
-export { DescriptionSection } from "./description-section";
-export { FinalSection } from "./final-section";
