+++ src/features/auth/components/auth-card.tsx
@@ -1,0 +1,20 @@
+import { ReactNode } from "react";
+
+type AuthCardProps = {
+  title: string;
+  subtitle: string;
+  children: ReactNode;
+};
+
+export function AuthCard({ title, subtitle, children }: AuthCardProps) {
+  return (
+    <div className="animate-appear">
+      <div className="text-center mb-8">
+        <h1 className="text-3xl font-bold text-zinc-900 dark:text-zinc-100 mb-2">{title}</h1>
+        <p className="text-zinc-600 dark:text-zinc-400">{subtitle}</p>
+      </div>
+
+      <div className="bg-white dark:bg-zinc-900 rounded-xl shadow-lg border border-zinc-200 dark:border-zinc-800 p-8">{children}</div>
+    </div>
+  );
+}
