+++ src/features/auth/index.ts
@@ -1,0 +1,6 @@
+export { AuthCard } from "./components/auth-card";
+export { FormInput } from "./components/form-input";
+export { useLogin } from "./hooks/useLogin";
+export { useRegister } from "./hooks/useRegister";
+export * from "./schemas/auth.schema";
+export * from "./types/auth.types";
