+++ src/features/home/index.ts
@@ -1,0 +1,3 @@
+export { BannerSection } from "./components/banner-section";
+export { DescriptionSection } from "./components/description-section";
+export { FinalSection } from "./components/final-section";
