+++ src/shared/icons/wcp-icon.tsx
