+++ src/shared/providers/layout-provider.tsx
@@ -1,0 +1,15 @@
+"use client";
+
+import { ThemeProvider } from "next-themes";
+import { NextIntlClientProvider } from "next-intl";
+
+export function Providers({ children, messages, locale }: { children: React.ReactNode; messages: Record<string, unknown>; locale: string }) {
+  const timeZone = Intl.DateTimeFormat().resolvedOptions().timeZone;
+  return (
+    <NextIntlClientProvider messages={messages} locale={locale} timeZone={timeZone}>
+      <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
+        {children}
+      </ThemeProvider>
+    </NextIntlClientProvider>
+  );
+}
